+++ lib_xcore_math/src/arch/ref/chunk.c
@@ -76,7 +76,7 @@ void chunk_q30_power_series(
 
     int32_t pow = 0x40000000;
     for(unsigned i = 0; i < term_count; i++){
-      acc = vlmacc32(acc, pow, coef[8*i]);
+      acc = (int32_t) vlmacc32(acc, pow, coef[8*i]);
       pow = vlmul32(pow, b[k]);
     }
